🐛 onesignal: fix onesignal web being imported on android
